Extract area list component
